download jar before building + adjust bash script and jar paths for validator

+++ support/validate_zugferd.sh
@@ -3,15 +3,23 @@
 mustang_version="2.14.2"
 mustang_jar="Mustang-CLI.jar"
 
+# add scriptdir as a variable to also check support/ for the jar as it's the case on CI pipeline runs
+scriptdir=$(realpath -s "$0")
+scriptdir="${scriptdir%/*}"
+
 if [ ! -f "$mustang_jar" ]; then
-	wget -O "$mustang_jar" "https://github.com/ZUGFeRD/mustangproject/releases/download/core-$mustang_version/Mustang-CLI-$mustang_version.jar"
+	if [ -f "$scriptdir/$mustang_jar" ]; then
+		mustang_jar="$scriptdir/$mustang_jar"
+	else
+		wget -O "$mustang_jar" "https://github.com/ZUGFeRD/mustangproject/releases/download/core-$mustang_version/Mustang-CLI-$mustang_version.jar"
+	fi
 fi
 
 validate_zugferd () {
 	for f in "$@"
 	do
 		if [[ -f "$f" ]]; then
-			java -Xmx1G -Dfile.encoding=UTF-8 -jar $mustang_jar --no-notices --action validate --source  "$f"
+			java -Xmx1G -Dfile.encoding=UTF-8 -jar "$mustang_jar" --no-notices --action validate --source  "$f"
 		fi
 	done
 }
